+++ config/default/core.entity_form_display.node.alert.default.yml
@@ -42,7 +42,7 @@ content:
       maxlength:
         maxlength_js: 255
         maxlength_js_label: 'Content limited to @limit characters, remaining: <strong>@remaining</strong>'
-        maxlength_js_enforce: false
+        maxlength_js_enforce: true
         maxlength_js_truncate_html: false
   field_severity:
     type: options_select
